fix: scan background task, SNMP test endpoint, route data cleanup
- Fix async scan execution: make trigger_scan async so asyncio.create_task works
- Add POST /api/switches/test for SNMP connectivity verification
- Add test-connection button to SwitchFormDialog
- Add auto-polling in SwitchDetail after scan trigger
- Fix SQLite BigInteger PK → Integer for autoincrement compatibility
- Fix route table garbled data via _bytes_to_ip() handling pysnmp OctetString
- Add robust exception handling in _run_scan_async
- Add string sanitation for SNMP values before DB storage

+++ backend/app/models/route_table.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, Integer, String, DateTime, ForeignKey, Index, func
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Index, func
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -10,7 +10,7 @@ class RouteTable(Base):
         Index("idx_rt_switch", "switch_id"),
     )
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     switch_id = Column(Integer, ForeignKey("switches.id", ondelete="CASCADE"), nullable=False)
     target_network = Column(String(45), nullable=False)
     subnet_mask = Column(String(45), nullable=False)
