+++ backend/app/models/scan_result.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, Integer, String, DateTime, ForeignKey, Index, func
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Index, func
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -12,7 +12,7 @@ class ScanResult(Base):
         Index("idx_sr_ip_mac", "ip_address", "mac_address"),
     )
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     switch_id = Column(Integer, ForeignKey("switches.id", ondelete="CASCADE"), nullable=False)
     ip_address = Column(String(45), nullable=False, default="")
     mac_address = Column(String(17), nullable=False)
